requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-WTF>=1.2
WTForms>=3.1
Werkzeug>=3.0
email-validator>=2.1
python-dotenv>=1.0
psycopg2-binary>=2.9
